Added few comments
