+++ siteManagerDashboard/utils.js
@@ -1,0 +1,16 @@
+
+
+export function getCurrentTimeStamp() {
+    const currentDate = new Date();
+    const currentMonth = currentDate.getMonth();
+    const currentDayOfMonth = currentDate.getDate();
+    const currentYear = currentDate.getFullYear();
+    const currentHour = currentDate.getHours();
+    const currentMinute = currentDate.getMinutes();
+    const currentSecond = currentDate.getSeconds();
+    const currentMillisecond = currentDate.getMilliseconds();
+    const timeStamp = currentYear + "-" + (currentMonth + 1)  + "-" + currentDayOfMonth + "T" 
+                        + currentHour + ":" + currentMinute + ":" + currentSecond + ":" + currentMillisecond;
+    
+    return timeStamp;
+}
